models/titanic_sample/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
numpy==1.13.1
scipy==0.19.1
scikit-learn==0.19.0
google-cloud-storage==1.8.0